+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K855G3ZHP8KYN7ARQ3R63PHQ"
+        "01K85KQYYG6YX4KYPVQEWBSW5Q"
     )
     r"""The id of the resource"""