+++ packages/orq-rc/src/orq_ai_sdk/models/updatebudgetop.py
@@ -208,5 +208,5 @@ class UpdateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-22T05:15:25.660Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-22T09:24:21.985Z")
     r"""The date and time the resource was last updated"""